pyproject.toml
[tool.poetry]
name = "sjfirebaseai"
version = "1.0.0"
description = ""
authors = ["Kenechukwu Akubue <kengoon19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pyjnius = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
